Split up SDN master/node/proxy/CNI code
