+++ pkg/sdn/master/doc.go
@@ -1,0 +1,2 @@
+// Package master contains the OpenShift SDN code that runs on the master
+package master
