+++ pkg/sdn/master/netid/allocator.go
