+++ pkg/sdn/master/netid/netid_test.go
