+++ pkg/sdn/node/doc.go
@@ -1,0 +1,2 @@
+// Package node contains the OpenShift SDN networking code that runs on the nodes
+package node
