+++ pkg/sdn/node/egress_network_policy.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"fmt"
