+++ pkg/sdn/node/multitenant.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"sync"
