+++ pkg/sdn/node/node.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"fmt"
@@ -13,7 +13,7 @@ import (
 
 	log "github.com/golang/glog"
 
-	"github.com/openshift/origin/pkg/sdn/plugin/cniserver"
+	"github.com/openshift/origin/pkg/sdn/node/cniserver"
 
 	osclient "github.com/openshift/origin/pkg/client"
 	"github.com/openshift/origin/pkg/sdn"
@@ -112,7 +112,7 @@ type OsdnNode struct {
 }
 
 // Called by higher layers to create the plugin SDN node instance
-func NewNodePlugin(c *OsdnNodeConfig) (sdn.NodeInterface, error) {
+func New(c *OsdnNodeConfig) (sdn.NodeInterface, error) {
 	var policy osdnPolicy
 	var pluginId int
 	var minOvsVersion string
