+++ pkg/sdn/node/ovscontroller_test.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"fmt"
