+++ pkg/sdn/node/sdn_controller.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"fmt"
