+++ pkg/sdn/node/subnets.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	log "github.com/golang/glog"
