+++ pkg/sdn/node/vnids_test.go
@@ -1,4 +1,4 @@
-package plugin
+package node
 
 import (
 	"testing"
