+++ pkg/sdn/proxy/doc.go
@@ -1,0 +1,2 @@
+// Package proxy contains the OpenShift SDN code that runs as part of the service proxy
+package proxy
