+++ pkg/sdn/proxy/proxy.go
@@ -1,4 +1,4 @@
-package plugin
+package proxy
 
 import (
 	"fmt"
@@ -60,7 +60,7 @@ type OsdnProxy struct {
 }
 
 // Called by higher layers to create the proxy plugin instance; only used by nodes
-func NewProxyPlugin(pluginName string, osClient *osclient.Client, kClient kclientset.Interface) (sdn.ProxyInterface, error) {
+func New(pluginName string, osClient *osclient.Client, kClient kclientset.Interface) (sdn.ProxyInterface, error) {
 	if !sdn.IsOpenShiftMultitenantNetworkPlugin(pluginName) {
 		return nil, nil
 	}
